backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
psycopg2
python-jose
passlib[bcrypt]
oauthlib
authlib
httpx
celery
redis
pytest
pytest-cov
python-dotenv
alembic
werkzeug
pytz
jose
email_validator
python-multipart
pyjwt
fastapi-limiter
aioredis
aiosmtplib